fix: pass correct type

+++ pkg/prove/proof.go
@@ -100,13 +100,13 @@ func txSharePosition(txs types.Txs, txIndex uint64) (startSharePos, endSharePos 
 	totalLen := 0
 	for i := uint64(0); i < txIndex; i++ {
 		txLen := len(txs[i])
-		totalLen += (shares.DelimLen(txLen) + txLen)
+		totalLen += (shares.DelimLen(uint64(txLen)) + txLen)
 	}
 
 	txLen := len(txs[txIndex])
 
 	startSharePos = uint64((totalLen) / appconsts.CompactShareContentSize)
-	endSharePos = uint64((totalLen + txLen + shares.DelimLen(txLen)) / appconsts.CompactShareContentSize)
+	endSharePos = uint64((totalLen + txLen + shares.DelimLen(uint64(txLen))) / appconsts.CompactShareContentSize)
 
 	return startSharePos, endSharePos, nil
 }
